backend/pyproject.toml
[tool.poetry]
name = "ollama-webui-api"
version = "0.1.0"
description = "backend api for local db storage"
authors = ["nulzo <nolanpgregory@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
djangorestframework = "^3.15.1"
python-dotenv = "^1.0.1"
django-cors-headers = "^4.3.1"
werkzeug = "^3.0.2"
django-extensions = "^3.2.3"
pillow = "^10.4.0"
ollama = "^0.3.2"
rich = "^13.8.0"
django-rest-auth = "^0.9.5"
openai = "^1.52.0"
uvicorn = "^0.32.0"
django-cryptography = "^1.1"
django = "^5.1.4"
chromadb = "^0.5.23"
pydot = "^3.0.4"

[tool.poetry.group.dev.dependencies]
black = "^24.10.0"
isort = "^5.13.2"
ruff = "^0.8.3"
pytest = "^8.2.0"
pylint = "^3.1.0"
django-stubs = "^5.0.0"
pytest-cov = "^5.0.0"
autopep8 = "^2.1.0"
mypy = "^1.10.0"
flake8 = "^7.0.0"
autoflake = "^2.3.1"
autobahn = "^23.6.2"
coverage = "^7.6.9"
bandit = "^1.8.0"
vulture = "^2.14"
pip-audit = "^2.7.3"
pyparsing = "^3.2.1"
pydot = "^3.0.4"
graphviz = "^0.20.3"

[tool.black]
line-length = 100
target-version = ['py39', 'py310', 'py311', 'py312']
extend-exclude = '''
(
    | settings.py
)
'''

[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "base.settings"
python_files = ["test_*.py", "*_test.py"]

[tool.ruff]
exclude = [".venv", "tests/", "scripts"]
line-length = 100

[tool.ruff.per-file-ignores]
"__init__.py" = ["E402"]
"settings.py" = ["E501"]

[tool.mypy]
ignore_missing_imports = "True"
plugins = ["mypy_django_plugin.main"]

[tool.django-stubs]
django_settings_module = "settings.settings"

[tool.autoflake]
imports = ["django", "requests", "urllib3"]

[tool.bandit]
exclude_dirs = [".venv", "tests"]
skips = ["B101"]

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
